+++ app/models/wallet_exchange.rb
@@ -1,5 +1,6 @@
 class WalletExchange < ApplicationRecord
   # Relationships
+  belongs_to :user
   belongs_to :exchange
   belongs_to :source_wallet, class_name: "Wallet"
   belongs_to :target_wallet, class_name: "Wallet"
