+++ db/schema.rb
@@ -67,6 +67,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_03_29_140332) do
     t.decimal "source_wallet_amount_after", precision: 30, scale: 12, null: false
     t.decimal "target_wallet_amount_before", precision: 30, scale: 12, null: false
     t.decimal "target_wallet_amount_after", precision: 30, scale: 12, null: false
+    t.bigint "user_id", null: false
     t.bigint "exchange_id", null: false
     t.bigint "source_wallet_id", null: false
     t.bigint "target_wallet_id", null: false
@@ -76,6 +77,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_03_29_140332) do
     t.index ["exchange_id"], name: "index_wallet_exchanges_on_exchange_id"
     t.index ["source_wallet_id"], name: "index_wallet_exchanges_on_source_wallet_id"
     t.index ["target_wallet_id"], name: "index_wallet_exchanges_on_target_wallet_id"
+    t.index ["user_id"], name: "index_wallet_exchanges_on_user_id"
   end
 
   create_table "wallets", force: :cascade do |t|
@@ -92,6 +94,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_03_29_140332) do
   add_foreign_key "exchanges", "currencies", column: "receive_currency_id"
   add_foreign_key "exchanges", "currencies", column: "send_currency_id"
   add_foreign_key "wallet_exchanges", "exchanges"
+  add_foreign_key "wallet_exchanges", "users"
   add_foreign_key "wallet_exchanges", "wallets", column: "source_wallet_id"
   add_foreign_key "wallet_exchanges", "wallets", column: "target_wallet_id"
   add_foreign_key "wallets", "currencies"
